+++ Plugins/app/CMakeLists.txt
@@ -2,3 +2,4 @@
 
 add_subdirectory(video2mat/)
 add_subdirectory(standard/)
+add_subdirectory(filters/)
