+++ Plugins/app/filters/CMakeLists.txt
@@ -1,0 +1,6 @@
+﻿project ("Kuafoo_App_Video2Mat")
+
+SET(OpenCV_DIR ${OpenCV_DIR_Env})
+FIND_PACKAGE(OpenCV REQUIRED)
+add_executable(filters Filters.cpp Filters.h main.cpp)
+target_link_libraries(filters ${OpenCV_LIBS} core)
