+++ Plugins/app/filters/Filters.cpp
@@ -1,0 +1,9 @@
+//
+// Created by xiaoyong on 2021/1/27.
+//
+
+#include "Filters.h"
+
+void Filters::run() {
+
+}
